FIX: Resolve compilation errors and static analysis warnings
- OpportunityManager.h:155: Fixed reinterpret_cast to static_cast for
  proper uintptr_t to uint32_t conversion (resolves C2440, E0171)
- EdgeCaseDetection.h:2284: Added static_cast<int> for size_t to int
  conversion (resolves C4267 warning)
- PredictionTelemetry.h:916: Added bounds check for sample_index before
  array access to prevent out-of-bounds read (resolves C6385)
- GeometricPrediction.h:1808: Added g_sdk NULL check before dereferencing
  in telemetry timestamp (resolves C6011)

All critical compilation errors resolved. Remaining warnings:
- MSB8028: Build configuration (non-code issue)
- C26815: False positive for intentional c_str() pattern in DannyPredPlugin.cpp

+++ EdgeCaseDetection.h
@@ -2280,7 +2280,7 @@ namespace EdgeCases
             // 15 samples (750ms) = medium confidence (3-4 cycles)
             // 30 samples (1500ms) = high confidence (5-7 cycles)
             float sample_confidence_multiplier = 1.0f;
-            int num_samples = history.velocity_samples.size();
+            int num_samples = static_cast<int>(history.velocity_samples.size());
 
             if (num_samples < MovementHistory::MIN_SAMPLES_FOR_PREDICTION)
             {
